+++ src/Equinox.EventStore/EventStore.fs
@@ -12,14 +12,14 @@ type Direction = Forward | Backward with
 
 module Log =
 
-    /// <summary>Name of Property used for <c>Event</c> in <c>LogEvent</c>s.</summary>
+    /// <summary>Name of Property used for <c>Metric</c> in <c>LogEvent</c>s.</summary>
     let [<Literal>] PropertyTag = "esEvt"
 
     [<NoEquality; NoComparison>]
     type Measurement = { stream : string; interval : StopwatchInterval; bytes : int; count : int }
 
     [<NoEquality; NoComparison>]
-    type Event =
+    type Metric =
         | WriteSuccess of Measurement
         | WriteConflict of Measurement
         | Slice of Direction * Measurement
@@ -48,7 +48,7 @@ module Log =
 
     /// Attach a property to the log context to hold the metrics
     // Sidestep Log.ForContext converting to a string; see https://github.com/serilog/serilog/issues/1124
-    let event (value : Event) (log : ILogger) =
+    let event (value : Metric) (log : ILogger) =
         let enrich (e : LogEvent) = e.AddPropertyIfAbsent(LogEventProperty(PropertyTag, ScalarValue(value)))
         log.ForContext({ new Serilog.Core.ILogEventEnricher with member _.Enrich(evt, _) = enrich evt })
 
@@ -81,9 +81,9 @@ module Log =
                 | :? ScalarValue as x -> Some x.Value
                 | _ -> None
 
-            let (|EsMetric|_|) (logEvent : LogEvent) : Event option =
+            let (|EsMetric|_|) (logEvent : LogEvent) : Metric option =
                 match logEvent.Properties.TryGetValue(PropertyTag) with
-                | true, SerilogScalar (:? Event as e) -> Some e
+                | true, SerilogScalar (:? Metric as e) -> Some e
                 | _ -> None
 
             type Counter =
@@ -230,7 +230,7 @@ module private Read =
         let reqMetric : Log.Measurement = { stream = streamName; interval = t; bytes = bytes; count = count}
         let batches = (events.Length - 1) / batchSize + 1
         let action = match direction with Direction.Forward -> "LoadF" | Direction.Backward -> "LoadB"
-        let evt = Log.Event.Batch (direction, batches, reqMetric)
+        let evt = Log.Metric.Batch (direction, batches, reqMetric)
         (log |> Log.prop "bytes" bytes |> Log.event evt).Information(
             "Ges{action:l} stream={stream} count={count}/{batches} version={version}",
             action, streamName, count, batches, version)